day07 缓存菜品1

+++ sky-server/src/main/java/com/sky/service/CategoryService.java
@@ -39,5 +39,5 @@ public interface CategoryService {
      * @param type
      * @return
      */
-    List<Category> getCategoryByType(long type);
+    List<Category> getCategoryByType(Integer type);
 }
